Remove unneccesary interface
